Update samples to SDK 2.6.1
This commit also modifies the .sln file for C# samples. Samples have
been reordered for consistency, and tabs instead of spaces are used
(formatting). Visual Studio only understands tabs for .sln files.
The reason we had to change is that (with spaces), every time one
builds the Solution, Visual Studio appends a new guid in the .sln file.
This is no longer the case when using tabs.

This commit also standardizes the AssemblyInfo.cs files.