Samples: Add loading projector image with Zivid API
Current samples (C++ and Python) load an image of arbitrary
resolution. OpenCV is then used to resize the image to match
the resolution of the Zivid camera projector. We don't use
OpenCV in C# samples. Therefore, to have load image in C#
samples for projecting images, we are doing the following.
In this commit, we are adding options to load an image
that matches the resolution of the camera projector, using
Zivid API.